+++ api/src/main/java/com/cms/controller/brand/BrandController.java
@@ -1,6 +1,7 @@
 package com.cms.controller.brand;
 
-import com.cms.controller.brand.dto.BrandDto;
+import com.cms.controller.brand.dto.BrandInfoResponse;
+import com.cms.controller.brand.dto.BrandSaveRequest;
 import com.cms.domain.Brand;
 import com.cms.service.BrandService;
 import lombok.RequiredArgsConstructor;
@@ -16,18 +17,18 @@ public class BrandController {
     private final BrandService brandService;
 
     @GetMapping("/{id}")
-    public BrandDto.BrandInfoResponse getBrandInfo(@PathVariable Long id) {
+    public BrandInfoResponse getBrandInfo(@PathVariable Long id) {
         Brand brand = brandService.getBrandInfo(id);
 //        brand 단의 내부 로직은 몰라야 하기 때문에 dto 에서 변환
-        return BrandDto.BrandInfoResponse.of(brand);
+        return BrandInfoResponse.of(brand);
     }
 
     @PostMapping("")
-    public BrandDto.BrandInfoResponse createBrand(@RequestBody BrandDto.BrandSaveRequest requestDto) {
+    public BrandInfoResponse createBrand(@RequestBody BrandSaveRequest requestDto) {
         log.info("request: {}", requestDto);
-        Brand newBrandInstance = BrandDto.BrandSaveRequest.toEntity(requestDto);
+        Brand newBrandInstance = BrandSaveRequest.toEntity(requestDto);
         Brand brand = brandService.createBrand(newBrandInstance);
-        return BrandDto.BrandInfoResponse.of(brand);
+        return BrandInfoResponse.of(brand);
     }
 
 }
