+++ build.gradle
@@ -43,7 +43,7 @@ subprojects {
         developmentOnly 'org.springframework.boot:spring-boot-devtools'
         compileOnly 'org.projectlombok:lombok'
         annotationProcessor 'org.springframework.boot:spring-boot-configuration-processor'
-//        runtimeOnly 'com.h2database:h2'
+        testImplementation 'com.h2database:h2'
         runtimeOnly 'com.mysql:mysql-connector-j'
         annotationProcessor 'org.projectlombok:lombok'
         testImplementation 'org.springframework.security:spring-security-test'
